pyproject.toml
[project]
name = "langrapgh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.32.0",
    "ipykernel>=6.30.1",
    "langchain-openai>=0.3.33",
    "langchain[anthropic]>=0.3.27",
    "langgraph>=0.6.7",
    "openai>=1.109.1",
    "python-dotenv>=1.1.1",
]
